Move simulate_spectrum to instrument
